Added unit tests for configuration projects.
Refactoring of the configuration API project. Unnecessary classes have been moved to their specific projects and the UserConfigurationService is now an interface that provider may implement. A default implementation with the exact behaviour of the class exists as part of the API project.

Updated the Configuration.Database project to use internal classes for database mapping with conversion between API classes and internal classes as well as encrypted storage of configuration values in the database.